Data.OpenApi

Description

Swagger™ is a project used to describe and document RESTful APIs.

The Swagger specification defines a set of files required to describe such an API. These files can then be used by the Swagger-UI project to display the API and Swagger-Codegen to generate clients in various languages. Additional utilities can also take advantage of the resulting files, such as testing tools.

How to use this library

This section explains how to use this library to work with Swagger specification.

Monoid instances

Virtually all types representing Swagger specification have Monoid instances. The Monoid type class provides two methods — mempty and mappend.

In this library you can use mempty for a default/empty value. For instance:

>>> BSL.putStrLn $ encodePretty (mempty :: OpenApi)
{
    "components": {},
    "info": {
        "title": "",
        "version": ""
    },
    "openapi": "3.0.0",
    "paths": {}
}

As you can see some spec properties (e.g. "version") are there even when the spec is empty. That is because these properties are actually required ones.

You should always override the default (empty) value for these properties, although it is not strictly necessary:

>>> BSL.putStrLn $ encodePretty mempty { _infoTitle = "Todo API", _infoVersion = "1.0" }
{
    "title": "Todo API",
    "version": "1.0"
}

You can merge two values using mappend or its infix version (<>):

>>> BSL.putStrLn $ encodePretty $ mempty { _infoTitle = "Todo API" } <> mempty { _infoVersion = "1.0" }
{
    "title": "Todo API",
    "version": "1.0"
}

Lenses and prisms

Note: if you're working with the optics library, take a look at Data.OpenApi.Optics.

Since Swagger has a fairly complex structure, lenses and prisms are used to work comfortably with it. In combination with Monoid instances, lenses make it fairly simple to construct/modify any part of the specification:

>>> :{
BSL.putStrLn $ encodePretty $ (mempty :: OpenApi)
  & components . schemas .~ IOHM.fromList [ ("User", mempty & type_ ?~ OpenApiString) ]
  & paths .~
    IOHM.fromList [ ("/user", mempty & get ?~ (mempty
        & at 200 ?~ ("OK" & _Inline.content.at "application/json" ?~ (mempty & schema ?~ Ref (Reference "User")))
        & at 404 ?~ "User info not found")) ]
:}
{
    "components": {
        "schemas": {
            "User": {
                "type": "string"
            }
        }
    },
    "info": {
        "title": "",
        "version": ""
    },
    "openapi": "3.0.0",
    "paths": {
        "/user": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "description": "User info not found"
                    }
                }
            }
        }
    }
}

In the snippet above we declare an API with a single path /user. This path provides method GET which produces application/json output. It should respond with code 200 and body specified by schema User which is defined in definitions property of swagger specification. Alternatively it may respond with code 404 meaning that user info is not found.

For convenience, swagger2 uses classy field lenses. It means that field accessor names can be overloaded for different types. One such common field is description. Many components of a Swagger specification can have descriptions, and you can use the same name for them:

>>> BSL.putStrLn $ encodePretty $ (mempty :: Response) & description .~ "No content"
{
    "description": "No content"
}
>>> :{
BSL.putStrLn $ encodePretty $ (mempty :: Schema)
  & type_       ?~ OpenApiBoolean
  & description ?~ "To be or not to be"
:}
{
    "description": "To be or not to be",
    "type": "boolean"
}

Additionally, to simplify working with Response, both Operation and Responses have direct access to it via at code. Example:

>>> :{
BSL.putStrLn $ encodePretty $ (mempty :: Operation)
  & at 404 ?~ "Not found"
:}
{
    "responses": {
        "404": {
            "description": "Not found"
        }
    }
}

You might've noticed that type_ has an extra underscore in its name compared to, say, description field accessor. This is because type is a keyword in Haskell. A few other field accessors are modified in this way:

in_, type_, default_ (as keywords);

maximum_, minimum_, head_ (as conflicting with Prelude);

enum_ (as conflicting with Control.Lens).

Schema specification

ParamSchema and Schema are the two core types for data model specification.

ParamSchema t specifies all the common properties, available for every data schema. The t parameter imposes some restrictions on type and items properties (see OpenApiType and OpenApiItems).

Schema is used for request and response bodies and allows specifying objects with properties in addition to what ParamSchema provides.

In most cases you will have a Haskell data type for which you would like to define a corresponding schema. To facilitate this use case swagger2 provides two classes for schema encoding. Both these classes provide means to encode types as Swagger schemas.

ToParamSchema is intended to be used for primitive API endpoint parameters, such as query parameters, headers and URL path pieces. Its corresponding value-encoding class is ToHttpApiData (from http-api-data package).

ToSchema is used for request and response bodies and mostly differ from primitive parameters by allowing objects/mappings in addition to primitive types and arrays. Its corresponding value-encoding class is ToJSON (from aeson package).

While lenses and prisms make it easy to define schemas, it might be that you don't need to: ToSchema and ToParamSchema classes both have default Generic-based implementations!

ToSchema default implementation is also aligned with ToJSON default implementation with the only difference being for sum encoding. ToJSON defaults sum encoding to defaultTaggedObject, while ToSchema defaults to something which corresponds to ObjectWithSingleField. This is due to defaultTaggedObject behavior being hard to specify in Swagger.

Here's an example showing ToJSON–ToSchema correspondance:

>>> data Person = Person { name :: String, age :: Integer } deriving Generic
>>> instance ToJSON Person
>>> instance ToSchema Person
>>> BSL.putStrLn $ encodePretty (Person "David" 28)
{
    "age": 28,
    "name": "David"
}
>>> BSL.putStrLn $ encodePretty $ toSchema (Proxy :: Proxy Person)
{
    "properties": {
        "age": {
            "type": "integer"
        },
        "name": {
            "type": "string"
        }
    },
    "required": [
        "name",
        "age"
    ],
    "type": "object"
}

This package implements OpenAPI 3.0 spec, which supports oneOf in schemas, allowing any sum types to be faithfully represented. All sum encodings supported by aeson are supported here as well, with an exception of TwoElemArray, since OpenAPI spec does not support heterogeneous arrays.

An example with TaggedObject encoding:

>>> data Error = ErrorNoUser { userId :: Int } | ErrorAccessDenied { requiredPermission :: String } deriving Generic
>>> instance ToJSON Error
>>> instance ToSchema Error
>>> BSL.putStrLn $ encodePretty $ toSchema (Proxy :: Proxy Error)
{
    "oneOf": [
        {
            "properties": {
                "tag": {
                    "enum": [
                        "ErrorNoUser"
                    ],
                    "type": "string"
                },
                "userId": {
                    "maximum": 9223372036854775807,
                    "minimum": -9223372036854775808,
                    "type": "integer"
                }
            },
            "required": [
                "userId",
                "tag"
            ],
            "title": "ErrorNoUser",
            "type": "object"
        },
        {
            "properties": {
                "requiredPermission": {
                    "type": "string"
                },
                "tag": {
                    "enum": [
                        "ErrorAccessDenied"
                    ],
                    "type": "string"
                }
            },
            "required": [
                "requiredPermission",
                "tag"
            ],
            "title": "ErrorAccessDenied",
            "type": "object"
        }
    ]
}

Manipulation

Sometimes you have to work with an imported or generated Swagger. For instance, http://hackage.haskell.org/package/servant-swagger generates basic Swagger for a type-level servant API.

Lenses and prisms can be used to manipulate such specification to add additional information, tags, extra responses, etc. To facilitate common needs, Data.OpenApi.Operation module provides useful helpers.

Validation

While ToParamSchema and ToSchema provide means to easily obtain schemas for Haskell types, there is no static mechanism to ensure those instances correspond to the ToHttpApiData or ToJSON instances.

Data.OpenApi.Schema.Validation addresses ToJSON/ToSchema validation.

Re-exports

module Data.OpenApi.Lens

module Data.OpenApi.Operation

module Data.OpenApi.ParamSchema

module Data.OpenApi.Schema

module Data.OpenApi.Schema.Validation

Swagger specification

data OpenApi

This is the root document object for the API specification.

Constructors

OpenApi

Fields

_openApiInfo :: Info

Provides metadata about the API. The metadata can be used by the clients if needed.

_openApiServers :: [Server]

An array of Server Objects, which provide connectivity information to a target server. If the servers property is not provided, or is an empty array, the default value would be a Server object with a url value of /.

_openApiPaths :: InsOrdHashMap FilePath PathItem

The available paths and operations for the API.

_openApiComponents :: Components

An element to hold various schemas for the specification.

_openApiSecurity :: [SecurityRequirement]

A declaration of which security mechanisms can be used across the API. The list of values includes alternative security requirement objects that can be used. Only one of the security requirement objects need to be satisfied to authorize a request. Individual operations can override this definition. To make security optional, an empty security requirement can be included in the array.

_openApiTags :: InsOrdHashSet Tag

A list of tags used by the specification with additional metadata. The order of the tags can be used to reflect on their order by the parsing tools. Not all tags that are used by the Operation Object must be declared. The tags that are not declared MAY be organized randomly or based on the tools' logic. Each tag name in the list MUST be unique.

_openApiExternalDocs :: Maybe ExternalDocs

Additional external documentation.

_openApiOpenapi :: OpenApiSpecVersion

The spec of OpenApi this spec adheres to. Must be between lowerOpenApiSpecVersion and upperOpenApiSpecVersion

data Server

An object representing a Server.

Constructors

Server

Fields

_serverUrl :: Text

A URL to the target host. This URL supports Server Variables and MAY be relative, to indicate that the host location is relative to the location where the OpenAPI document is being served. Variable substitutions will be made when a variable is named in {brackets}.

_serverDescription :: Maybe Text

An optional string describing the host designated by the URL. CommonMark syntax MAY be used for rich text representation.

_serverVariables :: InsOrdHashMap Text ServerVariable

A map between a variable name and its value. The value is used for substitution in the server's URL template.

data ServerVariable

Constructors

ServerVariable

Fields

_serverVariableEnum :: Maybe (InsOrdHashSet Text)

An enumeration of string values to be used if the substitution options are from a limited set. The array SHOULD NOT be empty.

_serverVariableDefault :: Text

The default value to use for substitution, which SHALL be sent if an alternate value is not supplied. Note this behavior is different than the 'Schema Object's treatment of default values, because in those cases parameter values are optional. If the _serverVariableEnum is defined, the value SHOULD exist in the enum's values.

_serverVariableDescription :: Maybe Text

An optional description for the server variable. CommonMark syntax MAY be used for rich text representation.

data Components

Holds a set of reusable objects for different aspects of the OAS. All objects defined within the components object will have no effect on the API unless they are explicitly referenced from properties outside the components object.

Constructors

Components

Fields

_componentsSchemas :: Definitions Schema

_componentsResponses :: Definitions Response

_componentsParameters :: Definitions Param

_componentsExamples :: Definitions Example

_componentsRequestBodies :: Definitions RequestBody

_componentsHeaders :: Definitions Header

_componentsSecuritySchemes :: SecurityDefinitions

_componentsLinks :: Definitions Link

_componentsCallbacks :: Definitions Callback

Info types

data Info

The object provides metadata about the API. The metadata MAY be used by the clients if needed, and MAY be presented in editing or documentation generation tools for convenience.

Constructors

Info

Fields

_infoTitle :: Text

The title of the API.

_infoDescription :: Maybe Text

A short description of the API. CommonMark syntax MAY be used for rich text representation.

_infoTermsOfService :: Maybe Text

A URL to the Terms of Service for the API. MUST be in the format of a URL.

_infoContact :: Maybe Contact

The contact information for the exposed API.

_infoLicense :: Maybe License

The license information for the exposed API.

_infoVersion :: Text

The version of the OpenAPI document (which is distinct from the OpenAPI Specification version or the API implementation version).

data Contact

Contact information for the exposed API.

Constructors

Contact

Fields

_contactName :: Maybe Text

The identifying name of the contact person/organization.

_contactUrl :: Maybe URL

The URL pointing to the contact information.

_contactEmail :: Maybe Text

The email address of the contact person/organization.

data License

License information for the exposed API.

Constructors

License

Fields

_licenseName :: Text

The license name used for the API.

_licenseUrl :: Maybe URL

A URL to the license used for the API.

PathItem

data PathItem

Describes the operations available on a single path. A PathItem may be empty, due to ACL constraints. The path itself is still exposed to the documentation viewer but they will not know which operations and parameters are available.

Constructors

PathItem

Fields

_pathItemSummary :: Maybe Text

An optional, string summary, intended to apply to all operations in this path.

_pathItemDescription :: Maybe Text

An optional, string description, intended to apply to all operations in this path. CommonMark syntax MAY be used for rich text representation.

_pathItemGet :: Maybe Operation

A definition of a GET operation on this path.

_pathItemPut :: Maybe Operation

A definition of a PUT operation on this path.

_pathItemPost :: Maybe Operation

A definition of a POST operation on this path.

_pathItemDelete :: Maybe Operation

A definition of a DELETE operation on this path.

_pathItemOptions :: Maybe Operation

A definition of a OPTIONS operation on this path.

_pathItemHead :: Maybe Operation

A definition of a HEAD operation on this path.

_pathItemPatch :: Maybe Operation

A definition of a PATCH operation on this path.

_pathItemTrace :: Maybe Operation

A definition of a TRACE operation on this path.

_pathItemServers :: [Server]

An alternative server array to service all operations in this path.

_pathItemParameters :: [Referenced Param]

A list of parameters that are applicable for all the operations described under this path. These parameters can be overridden at the operation level, but cannot be removed there. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location.

Operations

data Operation

Describes a single API operation on a path.

Constructors

Operation

Fields

_operationTags :: InsOrdHashSet TagName

A list of tags for API documentation control. Tags can be used for logical grouping of operations by resources or any other qualifier.

_operationSummary :: Maybe Text

A short summary of what the operation does. For maximum readability in the swagger-ui, this field SHOULD be less than 120 characters.

_operationDescription :: Maybe Text

A verbose explanation of the operation behavior. CommonMark syntax can be used for rich text representation.

_operationExternalDocs :: Maybe ExternalDocs

Additional external documentation for this operation.

_operationOperationId :: Maybe Text

Unique string used to identify the operation. The id MUST be unique among all operations described in the API. The operationId value is **case-sensitive**. Tools and libraries MAY use the operationId to uniquely identify an operation, therefore, it is RECOMMENDED to follow common programming naming conventions.

_operationParameters :: [Referenced Param]

A list of parameters that are applicable for this operation. If a parameter is already defined at the PathItem, the new definition will override it, but can never remove it. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location.

_operationRequestBody :: Maybe (Referenced RequestBody)

The request body applicable for this operation. The requestBody is only supported in HTTP methods where the HTTP 1.1 specification RFC7231 has explicitly defined semantics for request bodies. In other cases where the HTTP spec is vague, requestBody SHALL be ignored by consumers.

_operationResponses :: Responses

The list of possible responses as they are returned from executing this operation.

_operationCallbacks :: InsOrdHashMap Text (Referenced Callback)

A map of possible out-of band callbacks related to the parent operation. The key is a unique identifier for the Callback Object. Each value in the map is a Callback Object that describes a request that may be initiated by the API provider and the expected responses.

_operationDeprecated :: Maybe Bool

Declares this operation to be deprecated. Usage of the declared operation should be refrained. Default value is False.

_operationSecurity :: [SecurityRequirement]

A declaration of which security schemes are applied for this operation. The list of values describes alternative security schemes that can be used (that is, there is a logical OR between the security requirements). This definition overrides any declared top-level security. To remove a top-level security declaration, Just [] can be used.

_operationServers :: [Server]

An alternative server array to service this operation. If an alternative server object is specified at the PathItem Object or Root level, it will be overridden by this value.

data Tag

Allows adding meta data to a single tag that is used by Operation. It is not mandatory to have a Tag per tag used there.

Constructors

Tag

Fields

_tagName :: TagName

The name of the tag.

_tagDescription :: Maybe Text

A short description for the tag. CommonMark syntax MAY be used for rich text representation.

_tagExternalDocs :: Maybe ExternalDocs

Additional external documentation for this tag.

type TagName = Text

Tag name.

Types and formats

data OpenApiType where

Constructors

OpenApiString :: OpenApiType

OpenApiNumber :: OpenApiType

OpenApiInteger :: OpenApiType

OpenApiBoolean :: OpenApiType

OpenApiArray :: OpenApiType

OpenApiNull :: OpenApiType

OpenApiObject :: OpenApiType

type Format = Text

type Definitions = InsOrdHashMap Text

A list of definitions that can be used in references.

data Style

In order to support common ways of serializing simple parameters, a set of style values are defined.

Constructors

StyleMatrix
Path-style parameters defined by RFC6570.

StyleLabel
Label style parameters defined by RFC6570.

StyleForm
Form style parameters defined by RFC6570. This option replaces collectionFormat with a csv (when explode is false) or multi (when explode is true) value from OpenAPI 2.0.

StyleSimple
Simple style parameters defined by RFC6570. This option replaces collectionFormat with a csv value from OpenAPI 2.0.

StyleSpaceDelimited
Space separated array values. This option replaces collectionFormat equal to ssv from OpenAPI 2.0.

StylePipeDelimited
Pipe separated array values. This option replaces collectionFormat equal to pipes from OpenAPI 2.0.

StyleDeepObject
Provides a simple way of rendering nested objects using form parameters.

Parameters

data Param

Describes a single operation parameter. A unique parameter is defined by a combination of a name and location.

Constructors

Param

Fields

_paramName :: Text

The name of the parameter. Parameter names are case sensitive.

_paramDescription :: Maybe Text

A brief description of the parameter. This could contain examples of use. CommonMark syntax MAY be used for rich text representation.

_paramRequired :: Maybe Bool

Determines whether this parameter is mandatory. If the parameter is in "path", this property is required and its value MUST be true. Otherwise, the property MAY be included and its default value is False.

_paramDeprecated :: Maybe Bool

Specifies that a parameter is deprecated and SHOULD be transitioned out of usage. Default value is false.

_paramIn :: ParamLocation

The location of the parameter.

_paramAllowEmptyValue :: Maybe Bool

Sets the ability to pass empty-valued parameters. This is valid only for ParamQuery parameters and allows sending a parameter with an empty value. Default value is false.

_paramAllowReserved :: Maybe Bool

Determines whether the parameter value SHOULD allow reserved characters, as defined by RFC3986 :/?#[]!$&'()*+,;=@ to be included without percent-encoding. This property only applies to parameters with an _paramIn value of ParamQuery. The default value is False.

_paramSchema :: Maybe (Referenced Schema)

Parameter schema.

_paramStyle :: Maybe Style

Describes how the parameter value will be serialized depending on the type of the parameter value. Default values (based on value of _paramIn): for ParamQuery - StyleForm; for ParamPath - StyleSimple; for ParamHeader - StyleSimple; for ParamCookie - StyleForm.

_paramExplode :: Maybe Bool

When this is true, parameter values of type array or object generate separate parameters for each value of the array or key-value pair of the map. For other types of parameters this property has no effect. When style is form, the default value is true. For all other styles, the default value is false.

_paramExample :: Maybe Value

Example of the parameter's potential value. The example SHOULD match the specified schema and encoding properties if present. The _paramExample field is mutually exclusive of the _paramExamples field. Furthermore, if referencing a schema that contains an example, the example value SHALL override the example provided by the schema. To represent examples of media types that cannot naturally be represented in JSON or YAML, a string value can contain the example with escaping where necessary.

_paramExamples :: InsOrdHashMap Text (Referenced Example)

Examples of the parameter's potential value. Each example SHOULD contain a value in the correct format as specified in the parameter encoding. The _paramExamples field is mutually exclusive of the _paramExample field. Furthermore, if referencing a schema that contains an example, the examples value SHALL override the example provided by the schema.

data ParamLocation

Constructors

ParamQuery
Parameters that are appended to the URL. For example, in /items?id=###, the query parameter is id.

ParamHeader
Custom headers that are expected as part of the request.

ParamPath
Used together with Path Templating, where the parameter value is actually part of the operation's URL. This does not include the host or base path of the API. For example, in items{itemId}, the path parameter is itemId.

ParamCookie
Used to pass a specific cookie value to the API.

type ParamName = Text

data Header

Header fields have the same meaning as for Param.

Style is always treated as StyleSimple, as it is the only value allowed for headers.

Constructors

Header

Fields

_headerDescription :: Maybe HeaderName

A short description of the header.

_headerRequired :: Maybe Bool

_headerDeprecated :: Maybe Bool

_headerAllowEmptyValue :: Maybe Bool

_headerExplode :: Maybe Bool

_headerExample :: Maybe Value

_headerExamples :: InsOrdHashMap Text (Referenced Example)

_headerSchema :: Maybe (Referenced Schema)

type HeaderName = Text

data Example

Constructors

Example

Fields

_exampleSummary :: Maybe Text

Short description for the example.

_exampleDescription :: Maybe Text

Long description for the example. CommonMark syntax MAY be used for rich text representation.

_exampleValue :: Maybe Value

Embedded literal example. The _exampleValue field and _exampleExternalValue field are mutually exclusive.

To represent examples of media types that cannot naturally represented in JSON or YAML, use a string value to contain the example, escaping where necessary.

_exampleExternalValue :: Maybe URL

A URL that points to the literal example. This provides the capability to reference examples that cannot easily be included in JSON or YAML documents. The _exampleValue field and _exampleExternalValue field are mutually exclusive.

data RequestBody

Describes a single request body.

Constructors

RequestBody

Fields

_requestBodyDescription :: Maybe Text

A brief description of the request body. This could contain examples of use. CommonMark syntax MAY be used for rich text representation.

_requestBodyContent :: InsOrdHashMap MediaType MediaTypeObject

The content of the request body. The key is a media type or media type range and the value describes it. For requests that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text/*

_requestBodyRequired :: Maybe Bool

Determines if the request body is required in the request. Defaults to False.

data MediaTypeObject

Each Media Type Object provides schema and examples for the media type identified by its key.

Constructors

MediaTypeObject

Fields

_mediaTypeObjectSchema :: Maybe (Referenced Schema)

_mediaTypeObjectExample :: Maybe Value

Example of the media type. The example object SHOULD be in the correct format as specified by the media type.

_mediaTypeObjectExamples :: InsOrdHashMap Text (Referenced Example)

Examples of the media type. Each example object SHOULD match the media type and specified schema if present.

_mediaTypeObjectEncoding :: InsOrdHashMap Text Encoding

A map between a property name and its encoding information. The key, being the property name, MUST exist in the schema as a property. The encoding object SHALL only apply to RequestBody objects when the media type is multipart or application/x-www-form-urlencoded.

data Encoding

Constructors

Encoding

Fields

_encodingContentType :: Maybe MediaType

The Content-Type for encoding a specific property. Default value depends on the property type: for string with format being binary – application/octet-stream; for other primitive types – text/plain; for object - application/json; for array – the default is defined based on the inner type. The value can be a specific media type (e.g. application/json), a wildcard media type (e.g. image/*), or a comma-separated list of the two types.

_encodingHeaders :: InsOrdHashMap Text (Referenced Header)

A map allowing additional information to be provided as headers, for example Content-Disposition. Content-Type is described separately and SHALL be ignored in this section. This property SHALL be ignored if the request body media type is not a multipart.

_encodingStyle :: Maybe Style

Describes how a specific property value will be serialized depending on its type. See Param Object for details on the style property. The behavior follows the same values as query parameters, including default values. This property SHALL be ignored if the request body media type is not application/x-www-form-urlencoded.

_encodingExplode :: Maybe Bool

When this is true, property values of type array or object generate separate parameters for each value of the array, or key-value-pair of the map. For other types of properties this property has no effect. When style is form, the default value is true. For all other styles, the default value is false. This property SHALL be ignored if the request body media type is not application/x-www-form-urlencoded.

_encodingAllowReserved :: Maybe Bool

Determines whether the parameter value SHOULD allow reserved characters, as defined by RFC3986 :/?#[]!$&'()*+,;= to be included without percent-encoding. The default value is false. This property SHALL be ignored if the request body media type is not application/x-www-form-urlencoded@.

Schemas

data Schema

Constructors

Schema

Fields

_schemaTitle :: Maybe Text

_schemaDescription :: Maybe Text

_schemaRequired :: [ParamName]

_schemaNullable :: Maybe Bool

_schemaAllOf :: Maybe [Referenced Schema]

_schemaOneOf :: Maybe [Referenced Schema]

_schemaNot :: Maybe (Referenced Schema)

_schemaAnyOf :: Maybe [Referenced Schema]

_schemaProperties :: InsOrdHashMap Text (Referenced Schema)

_schemaAdditionalProperties :: Maybe AdditionalProperties

_schemaDiscriminator :: Maybe Discriminator

_schemaReadOnly :: Maybe Bool

_schemaWriteOnly :: Maybe Bool

_schemaXml :: Maybe Xml

_schemaExternalDocs :: Maybe ExternalDocs

_schemaExample :: Maybe Value

_schemaDeprecated :: Maybe Bool

_schemaMaxProperties :: Maybe Integer

_schemaMinProperties :: Maybe Integer

_schemaDefault :: Maybe Value

Declares the value of the parameter that the server will use if none is provided, for example a "count" to control the number of results per page might default to 100 if not supplied by the client in the request. (Note: "default" has no meaning for required parameters.) Unlike JSON Schema this value MUST conform to the defined type for this parameter.

_schemaType :: Maybe OpenApiType

_schemaFormat :: Maybe Format

_schemaItems :: Maybe OpenApiItems

_schemaMaximum :: Maybe Scientific

_schemaExclusiveMaximum :: Maybe Bool

_schemaMinimum :: Maybe Scientific

_schemaExclusiveMinimum :: Maybe Bool

_schemaMaxLength :: Maybe Integer

_schemaMinLength :: Maybe Integer

_schemaPattern :: Maybe Pattern

_schemaMaxItems :: Maybe Integer

_schemaMinItems :: Maybe Integer

_schemaUniqueItems :: Maybe Bool

_schemaEnum :: Maybe [Value]

_schemaMultipleOf :: Maybe Scientific

data NamedSchema

A Schema with an optional name. This name can be used in references.

Constructors

NamedSchema

Fields

_namedSchemaName :: Maybe Text

_namedSchemaSchema :: Schema

data OpenApiItems where

Items for OpenApiArray schemas.

Warning: OpenAPI 3.0 does not support tuple arrays. However, OpenAPI 3.1 will, as it will incorporate Json Schema mostly verbatim.

OpenApiItemsObject should be used to specify homogenous array Schemas.

OpenApiItemsArray should be used to specify tuple Schemas.

Constructors

OpenApiItemsObject :: Referenced Schema -> OpenApiItems

OpenApiItemsArray :: [Referenced Schema] -> OpenApiItems

data Xml

Constructors

Xml

Fields

_xmlName :: Maybe Text

Replaces the name of the element/attribute used for the described schema property. When defined within the OpenApiItems (items), it will affect the name of the individual XML elements within the list. When defined alongside type being array (outside the items), it will affect the wrapping element and only if wrapped is true. If wrapped is false, it will be ignored.

_xmlNamespace :: Maybe Text

The URL of the namespace definition. Value SHOULD be in the form of a URL.

_xmlPrefix :: Maybe Text

The prefix to be used for the name.

_xmlAttribute :: Maybe Bool

Declares whether the property definition translates to an attribute instead of an element. Default value is False.

_xmlWrapped :: Maybe Bool

MAY be used only for an array definition. Signifies whether the array is wrapped (for example, <books><book><book></books>) or unwrapped (<book><book>). Default value is False. The definition takes effect only when defined alongside type being array (outside the items).

type Pattern = Text

Regex pattern for string type.

data AdditionalProperties

Constructors

AdditionalPropertiesAllowed Bool

AdditionalPropertiesSchema (Referenced Schema)

data Discriminator

Constructors

Discriminator

Fields

_discriminatorPropertyName :: Text

The name of the property in the payload that will hold the discriminator value.

_discriminatorMapping :: InsOrdHashMap Text Text

An object to hold mappings between payload values and schema names or references.

Responses

data Responses

A container for the expected responses of an operation. The container maps a HTTP response code to the expected response. It is not expected from the documentation to necessarily cover all possible HTTP response codes, since they may not be known in advance. However, it is expected from the documentation to cover a successful operation response and any known errors.

Constructors

Responses

Fields

_responsesDefault :: Maybe (Referenced Response)

The documentation of responses other than the ones declared for specific HTTP response codes. It can be used to cover undeclared responses.

_responsesResponses :: InsOrdHashMap HttpStatusCode (Referenced Response)

Any HTTP status code can be used as the property name (one property per HTTP status code). Describes the expected response for those HTTP status codes.

data Response

Describes a single response from an API Operation.

Constructors

Response

Fields

_responseDescription :: Text

A short description of the response. CommonMark syntax can be used for rich text representation.

_responseContent :: InsOrdHashMap MediaType MediaTypeObject

A map containing descriptions of potential response payloads. The key is a media type or media type range and the value describes it. For responses that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text/*.

_responseHeaders :: InsOrdHashMap HeaderName (Referenced Header)

Maps a header name to its definition.

_responseLinks :: InsOrdHashMap Text (Referenced Link)

A map of operations links that can be followed from the response. The key of the map is a short name for the link, following the naming constraints of the names for Component Objects.

type HttpStatusCode = Int

data Link

The Link object represents a possible design-time link for a response. The presence of a link does not guarantee the caller's ability to successfully invoke it, rather it provides a known relationship and traversal mechanism between responses and other operations.

Constructors

Link

Fields

_linkOperationRef :: Maybe Text

A relative or absolute URI reference to an OAS operation. This field is mutually exclusive of the _linkOperationId field, and MUST point to an Operation Object. Relative _linkOperationRef values MAY be used to locate an existing Operation Object in the OpenAPI definition.

_linkOperationId :: Maybe Text

The name of an existing, resolvable OAS operation, as defined with a unique _operationOperationId. This field is mutually exclusive of the _linkOperationRef field.

_linkParameters :: InsOrdHashMap Text ExpressionOrValue

A map representing parameters to pass to an operation as specified with _linkOperationId or identified via _linkOperationRef. The key is the parameter name to be used, whereas the value can be a constant or an expression to be evaluated and passed to the linked operation. The parameter name can be qualified using the parameter location [{in}.]{name} for operations that use the same parameter name in different locations (e.g. path.id).

_linkRequestBody :: Maybe ExpressionOrValue

A literal value or {expression} to use as a request body when calling the target operation.

_linkDescription :: Maybe Text

A description of the link.

_linkServer :: Maybe Server

A server object to be used by the target operation.

newtype Callback

A map of possible out-of band callbacks related to the parent operation. Each value in the map is a PathItem Object that describes a set of requests that may be initiated by the API provider and the expected responses. The key value used to identify the path item object is an expression, evaluated at runtime, that identifies a URL to use for the callback operation.

Constructors

Callback (InsOrdHashMap Text PathItem)

Security

data SecurityScheme

Constructors

SecurityScheme

Fields

_securitySchemeType :: SecuritySchemeType

The type of the security scheme.

_securitySchemeDescription :: Maybe Text

A short description for security scheme.

data SecuritySchemeType

>>> BSL.putStrLn $ encodePretty (SecuritySchemeHttp (HttpSchemeBearer Nothing))
{
    "scheme": "bearer",
    "type": "http"
}

>>> BSL.putStrLn $ encodePretty (SecuritySchemeHttp (HttpSchemeBearer (Just "jwt")))
{
    "bearerFormat": "jwt",
    "scheme": "bearer",
    "type": "http"
}

>>> BSL.putStrLn $ encodePretty (SecuritySchemeHttp HttpSchemeBasic)
{
    "scheme": "basic",
    "type": "http"
}

>>> BSL.putStrLn $ encodePretty (SecuritySchemeHttp (HttpSchemeCustom "CANARY"))
{
    "scheme": "CANARY",
    "type": "http"
}

>>> BSL.putStrLn $ encodePretty (SecuritySchemeApiKey (ApiKeyParams "id" ApiKeyCookie))
{
    "in": "cookie",
    "name": "id",
    "type": "apiKey"
}

Constructors

SecuritySchemeHttp HttpSchemeType

SecuritySchemeApiKey ApiKeyParams

SecuritySchemeOAuth2 OAuth2Flows

SecuritySchemeOpenIdConnect URL

data HttpSchemeType

Constructors

HttpSchemeBearer (Maybe BearerFormat)

HttpSchemeBasic

HttpSchemeCustom Text

newtype SecurityDefinitions

Constructors

SecurityDefinitions (Definitions SecurityScheme)

newtype SecurityRequirement

Lists the required security schemes to execute this operation. The object can have multiple security schemes declared in it which are all required (that is, there is a logical AND between the schemes).

Constructors

SecurityRequirement

Fields

getSecurityRequirement :: InsOrdHashMap Text [Text]

API key

data ApiKeyParams

Constructors

ApiKeyParams

Fields

_apiKeyName :: Text

The name of the header or query parameter to be used.

_apiKeyIn :: ApiKeyLocation

The location of the API key.

data ApiKeyLocation

The location of the API key.

Constructors

ApiKeyQuery

ApiKeyHeader

ApiKeyCookie

OAuth2

data OAuth2Flows

Constructors

OAuth2Flows

Fields

_oAuth2FlowsImplicit :: Maybe (OAuth2Flow OAuth2ImplicitFlow)

Configuration for the OAuth Implicit flow

_oAuth2FlowsPassword :: Maybe (OAuth2Flow OAuth2PasswordFlow)

Configuration for the OAuth Resource Owner Password flow

_oAuth2FlowsClientCredentials :: Maybe (OAuth2Flow OAuth2ClientCredentialsFlow)

Configuration for the OAuth Client Credentials flow

_oAuth2FlowsAuthorizationCode :: Maybe (OAuth2Flow OAuth2AuthorizationCodeFlow)

Configuration for the OAuth Authorization Code flow

data OAuth2Flow p

Constructors

OAuth2Flow

Fields

_oAuth2Params :: p

_oAath2RefreshUrl :: Maybe URL

The URL to be used for obtaining refresh tokens.

_oAuth2Scopes :: InsOrdHashMap Text Text

The available scopes for the OAuth2 security scheme. A map between the scope name and a short description for it. The map MAY be empty.

newtype OAuth2ImplicitFlow

Constructors

OAuth2ImplicitFlow

Fields

_oAuth2ImplicitFlowAuthorizationUrl :: AuthorizationURL

newtype OAuth2PasswordFlow

Constructors

OAuth2PasswordFlow

Fields

_oAuth2PasswordFlowTokenUrl :: TokenURL

newtype OAuth2ClientCredentialsFlow

Constructors

OAuth2ClientCredentialsFlow

Fields

_oAuth2ClientCredentialsFlowTokenUrl :: TokenURL

data OAuth2AuthorizationCodeFlow

Constructors

OAuth2AuthorizationCodeFlow

Fields

_oAuth2AuthorizationCodeFlowAuthorizationUrl :: AuthorizationURL

_oAuth2AuthorizationCodeFlowTokenUrl :: TokenURL

type AuthorizationURL = Text

The authorization URL to be used for OAuth2 flow. This SHOULD be in the form of a URL.

type TokenURL = Text

The token URL to be used for OAuth2 flow. This SHOULD be in the form of a URL.

External documentation

data ExternalDocs

Allows referencing an external resource for extended documentation.

Constructors

ExternalDocs

Fields

_externalDocsDescription :: Maybe Text

A short description of the target documentation. CommonMark syntax MAY be used for rich text representation.

_externalDocsUrl :: URL

The URL for the target documentation.

References

newtype Reference

A simple object to allow referencing other definitions in the specification. It can be used to reference parameters and responses that are defined at the top level for reuse.

Constructors

Reference

Fields

getReference :: Text

data Referenced a

Constructors

Ref Reference

Inline a

Miscellaneous

newtype MimeList

Constructors

MimeList

Fields

getMimeList :: [MediaType]

newtype URL

Constructors

URL

Fields

getUrl :: Text