Data.OpenApi.Internal.ParamSchema

Documentation

binarySchema :: Schema

Default schema for binary data (any sequence of octets).

byteSchema :: Schema

Default schema for binary data (base64 encoded).

passwordSchema :: Schema

Default schema for password string. "password" format is used to hint UIs the input needs to be obscured.

class ToParamSchema a where

Convert a type into a plain Schema.

In previous versions of the package there was a separate type called ParamSchema, which was included in a greater Schema. Now this is a single class, but distinction for schema generators for "simple" types is preserved.

ToParamSchema is suited only for primitive-like types without nested fields and such.

An example type and instance:

{-# LANGUAGE OverloadedStrings #-}   -- allows to write Text literals

import Control.Lens

data Direction = Up | Down

instance ToParamSchema Direction where
  toParamSchema _ = mempty
     & type_ ?~ OpenApiString
     & enum_ ?~ [ "Up", "Down" ]

Instead of manually writing your ToParamSchema instance you can use a default generic implementation of toParamSchema.

To do that, simply add deriving FPFormat clause to your datatype and declare a ToParamSchema instance for your datatype without giving definition for toParamSchema.

For instance, the previous example can be simplified into this:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics (Generic)

data Direction = Up | Down deriving Generic

instance ToParamSchema Direction

Minimal complete definition

Nothing

Methods

toParamSchema :: Proxy a -> Schema

Convert a type into a plain parameter schema.

>>> BSL.putStrLn $ encodePretty $ toParamSchema (Proxy :: Proxy Integer)
{
    "type": "integer"
}

default toParamSchema :: (Generic a, GToParamSchema (Rep a)) => Proxy a -> Schema

toParamSchemaBoundedIntegral :: forall a t. (Bounded a, Integral a) => Proxy a -> Schema

Default plain schema for Bounded, Integral types.

>>> BSL.putStrLn $ encodePretty $ toParamSchemaBoundedIntegral (Proxy :: Proxy Int8)
{
    "maximum": 127,
    "minimum": -128,
    "type": "integer"
}

timeParamSchema :: String -> Schema

type family ToParamSchemaByteStringError bs where ...

Equations

ToParamSchemaByteStringError bs = TypeError (((('Text "Impossible to have an instance " :<>: ShowType (ToParamSchema bs)) :<>: Text ".") :$$: (('Text "Please, use a newtype wrapper around " :<>: ShowType bs) :<>: Text " instead.")) :$$: 'Text "Consider using byteParamSchema or binaryParamSchemaemplates.")

genericToParamSchema :: forall a t. (Generic a, GToParamSchema (Rep a)) => SchemaOptions -> Proxy a -> Schema

A configurable generic Schema creator.

>>> :set -XDeriveGeneric
>>> data Color = Red | Blue deriving Generic
>>> BSL.putStrLn $ encodePretty $ genericToParamSchema defaultSchemaOptions (Proxy :: Proxy Color)
{
    "enum": [
        "Red",
        "Blue"
    ],
    "type": "string"
}

class GToParamSchema (f :: * -> *) where

Methods

gtoParamSchema :: SchemaOptions -> Proxy f -> Schema -> Schema

class GEnumParamSchema (f :: * -> *) where

Methods

genumParamSchema :: SchemaOptions -> Proxy f -> Schema -> Schema

data Proxy3 a b c

Constructors

Proxy3

>>> import Data.Aeson (encode)
>>> import Data.OpenApi.Internal.Utils