Data.OpenApi.Internal.Schema.Validation

Description

Validate JSON values with Swagger Schema.

Documentation

>>> import Data.OpenApi.Internal.Schema.Validation

validatePrettyToJSON :: forall a. (ToJSON a, ToSchema a) => a -> Maybe String

Validate ToJSON instance matches ToSchema for a given value. This can be used with QuickCheck to ensure those instances are coherent:

validateToJSON (x :: Int) == []

NOTE: validateToJSON does not perform string pattern validation. See validateToJSONWithPatternChecker.

See renderValidationErrors on how the output is structured.

validateToJSON :: forall a. (ToJSON a, ToSchema a) => a -> [ValidationError]

Variant of validatePrettyToJSON with typed output.

validateToJSONWithPatternChecker :: forall a. (ToJSON a, ToSchema a) => (Pattern -> Text -> Bool) -> a -> [ValidationError]

Validate ToJSON instance matches ToSchema for a given value and pattern checker. This can be used with QuickCheck to ensure those instances are coherent.

For validation without patterns see validateToJSON. See also: renderValidationErrors.

renderValidationErrors :: forall a. (ToJSON a, ToSchema a) => (a -> [ValidationError]) -> a -> Maybe String

Pretty print validation errors together with actual JSON and Swagger Schema (using encodePretty).

>>> import Data.Aeson as Aeson
>>> import Data.Foldable (traverse_)
>>> import GHC.Generics
>>> data Phone = Phone { value :: String } deriving (Generic)
>>> data Person = Person { name :: String, phone :: Phone } deriving (Generic)
>>> instance ToJSON Person where toJSON p = object [ "name" Aeson..= name p ]
>>> instance ToSchema Phone
>>> instance ToSchema Person
>>> let person = Person { name = "John", phone = Phone "123456" }
>>> traverse_ putStrLn $ renderValidationErrors validateToJSON person
Validation against the schema fails:
  * property "phone" is required, but not found in "{\"name\":\"John\"}"

JSON value:
{
    "name": "John"
}

Swagger Schema:
{
    "properties": {
        "name": {
            "type": "string"
        },
        "phone": {
            "$ref": "#/components/schemas/Phone"
        }
    },
    "required": [
        "name",
        "phone"
    ],
    "type": "object"
}

Swagger Description Context:
{
    "Phone": {
        "properties": {
            "value": {
                "type": "string"
            }
        },
        "required": [
            "value"
        ],
        "type": "object"
    }
}

validateJSON :: Definitions Schema -> Schema -> Value -> [ValidationError]

Validate JSON ExpressionOrValue against Swagger Schema.

validateJSON mempty (toSchema (Proxy :: Proxy Int)) (toJSON (x :: Int)) == []

NOTE: validateJSON does not perform string pattern validation. See validateJSONWithPatternChecker.

validateJSONWithPatternChecker :: (Pattern -> Text -> Bool) -> Definitions Schema -> Schema -> Value -> [ValidationError]

Validate JSON ExpressionOrValue agains Swagger ToSchema for a given value and pattern checker.

For validation without patterns see validateJSON.

type ValidationError = String

Validation error message.

data Result a

Validation result type.

Constructors

Failed [ValidationError]
Validation failed with a list of error messages.

Passed a
Validation passed.

data Config

Validation configuration.

Constructors

Config

Fields

configPatternChecker :: Pattern -> Text -> Bool

Pattern checker for _schemaPattern validation.

configDefinitions :: Definitions Schema

Schema definitions in scope to resolve references.

defaultConfig :: Config

Default Config:

defaultConfig = Config
  { configPatternChecker = \_pattern _str -> True
  , configDefinitions    = mempty
  }

newtype Validation s a

Value validation.

Constructors

Validation

Fields

runValidation :: Config -> s -> Result a

withConfig :: (Config -> Validation s a) -> Validation s a

withSchema :: (s -> Validation s a) -> Validation s a

invalid :: String -> Validation schema a

Issue an error message.

valid :: Validation schema ()

Validation passed.

checkMissing :: Validation s () -> Lens' s (Maybe a) -> (a -> Validation s ()) -> Validation s ()

Validate schema's property given a lens into that property and property checker.

check :: Lens' s (Maybe a) -> (a -> Validation s ()) -> Validation s ()

Validate schema's property given a lens into that property and property checker. If property is missing in schema, consider it valid.

sub :: t -> Validation t a -> Validation s a

Validate same value with different schema.

sub_ :: Getting a s a -> Validation a r -> Validation s r

Validate same value with a part of the original schema.

withRef :: Reference -> (Schema -> Validation s a) -> Validation s a

Validate value against a schema given schema reference and validation function.

validateWithSchemaRef :: Referenced Schema -> Value -> Validation s ()

validateWithSchema :: Value -> Validation Schema ()

Validate JSON ExpressionOrValue with Swagger Schema.

validateInteger :: Scientific -> Validation Schema ()

validateNumber :: Scientific -> Validation Schema ()

validateString :: Text -> Validation Schema ()

validateArray :: Vector Value -> Validation Schema ()

validateObject :: KeyMap Value -> Validation Schema ()

validateEnum :: Value -> Validation Schema ()

inferSchemaTypes :: Schema -> [OpenApiType]

Infer schema type based on used properties.

This is like inferParamSchemaTypes, but also works for objects:

>>> inferSchemaTypes <$> decode "{\"minProperties\": 1}"
Just [OpenApiObject]

inferParamSchemaTypes :: Schema -> [OpenApiType]

Infer schema type based on used properties.

>>> inferSchemaTypes <$> decode "{\"minLength\": 2}"
Just [OpenApiString]

>>> inferSchemaTypes <$> decode "{\"maxItems\": 0}"
Just [OpenApiArray]

From numeric properties OpenApiInteger type is inferred. If you want OpenApiNumber instead, you must specify it explicitly.

>>> inferSchemaTypes <$> decode "{\"minimum\": 1}"
Just [OpenApiInteger]

validateSchemaType :: Value -> Validation Schema ()

validateParamSchemaType :: Value -> Validation Schema ()

showType :: (Maybe OpenApiType, Value) -> String