Data.OpenApi.ParamSchema

Description

Types and functions for working with Swagger parameter schema.

Encoding

class ToParamSchema a where

Convert a type into a plain Schema.

In previous versions of the package there was a separate type called ParamSchema, which was included in a greater Schema. Now this is a single class, but distinction for schema generators for "simple" types is preserved.

ToParamSchema is suited only for primitive-like types without nested fields and such.

An example type and instance:

{-# LANGUAGE OverloadedStrings #-}   -- allows to write Text literals

import Control.Lens

data Direction = Up | Down

instance ToParamSchema Direction where
  toParamSchema _ = mempty
     & type_ ?~ OpenApiString
     & enum_ ?~ [ "Up", "Down" ]

Instead of manually writing your ToParamSchema instance you can use a default generic implementation of toParamSchema.

To do that, simply add deriving FPFormat clause to your datatype and declare a ToParamSchema instance for your datatype without giving definition for toParamSchema.

For instance, the previous example can be simplified into this:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics (Generic)

data Direction = Up | Down deriving Generic

instance ToParamSchema Direction

Minimal complete definition

Nothing

Methods

toParamSchema :: Proxy a -> Schema

Convert a type into a plain parameter schema.

>>> BSL.putStrLn $ encodePretty $ toParamSchema (Proxy :: Proxy Integer)
{
    "type": "integer"
}

default toParamSchema :: (Generic a, GToParamSchema (Rep a)) => Proxy a -> Schema

Generic schema encoding

genericToParamSchema :: forall a t. (Generic a, GToParamSchema (Rep a)) => SchemaOptions -> Proxy a -> Schema

A configurable generic Schema creator.

>>> :set -XDeriveGeneric
>>> data Color = Red | Blue deriving Generic
>>> BSL.putStrLn $ encodePretty $ genericToParamSchema defaultSchemaOptions (Proxy :: Proxy Color)
{
    "enum": [
        "Red",
        "Blue"
    ],
    "type": "string"
}

toParamSchemaBoundedIntegral :: forall a t. (Bounded a, Integral a) => Proxy a -> Schema

Default plain schema for Bounded, Integral types.

>>> BSL.putStrLn $ encodePretty $ toParamSchemaBoundedIntegral (Proxy :: Proxy Int8)
{
    "maximum": 127,
    "minimum": -128,
    "type": "integer"
}

Schema templates

passwordSchema :: Schema

Default schema for password string. "password" format is used to hint UIs the input needs to be obscured.

binarySchema :: Schema

Default schema for binary data (any sequence of octets).

byteSchema :: Schema

Default schema for binary data (base64 encoded).

Generic encoding configuration

data SchemaOptions

Options that specify how to encode your type to Swagger schema.

Constructors

SchemaOptions

Fields

fieldLabelModifier :: String -> String

Function applied to field labels. Handy for removing common record prefixes for example.

constructorTagModifier :: String -> String

Function applied to constructor tags which could be handy for lower-casing them for example.

datatypeNameModifier :: String -> String

Function applied to datatype name.

allNullaryToStringTag :: Bool

If True the constructors of a datatype, with all nullary constructors, will be encoded to a string enumeration schema with the constructor tags as possible values.

unwrapUnaryRecords :: Bool

Hide the field name when a record constructor has only one field, like a newtype.

sumEncoding :: SumEncoding

Specifies how to encode constructors of a sum datatype.

defaultSchemaOptions :: SchemaOptions

Default encoding SchemaOptions.

SchemaOptions
{ fieldLabelModifier     = id
, constructorTagModifier = id
, datatypeNameModifier   = id
, allNullaryToStringTag  = True
, unwrapUnaryRecords     = False
, sumEncoding            = defaultTaggedObject
}