Data.OpenApi.SchemaOptions

Description

Generic deriving options for ToParamSchema and ToSchema.

Documentation

data SchemaOptions

Options that specify how to encode your type to Swagger schema.

Constructors

SchemaOptions

Fields

fieldLabelModifier :: String -> String

Function applied to field labels. Handy for removing common record prefixes for example.

constructorTagModifier :: String -> String

Function applied to constructor tags which could be handy for lower-casing them for example.

datatypeNameModifier :: String -> String

Function applied to datatype name.

allNullaryToStringTag :: Bool

If True the constructors of a datatype, with all nullary constructors, will be encoded to a string enumeration schema with the constructor tags as possible values.

unwrapUnaryRecords :: Bool

Hide the field name when a record constructor has only one field, like a newtype.

sumEncoding :: SumEncoding

Specifies how to encode constructors of a sum datatype.

defaultSchemaOptions :: SchemaOptions

Default encoding SchemaOptions.

SchemaOptions
{ fieldLabelModifier     = id
, constructorTagModifier = id
, datatypeNameModifier   = id
, allNullaryToStringTag  = True
, unwrapUnaryRecords     = False
, sumEncoding            = defaultTaggedObject
}

fromAesonOptions :: Options -> SchemaOptions

Convert Options to SchemaOptions.

Specifically the following fields get copied:

fieldLabelModifier

constructorTagModifier

allNullaryToStringTag

unwrapUnaryRecords

Note that these fields have no effect on SchemaOptions:

omitNothingFields

tagSingleConstructors

The rest is defined as in defaultSchemaOptions.

Since: 2.2.1